Netsh commands are in their own module now. Closes #7
