+++ src/renderer/components/NetworkInterfaces.vue
@@ -64,6 +64,8 @@
 
 <script>
     const os = require('os');
+    const netsh = require('../ipconfig/windows/netsh.js').default;
+    console.log(netsh);
     const EventEmitter = require('events').EventEmitter;
     import {detailedDiff} from 'deep-object-diff';
 
@@ -78,63 +80,6 @@
         console.log('updated event', event)
     });
 
-    /*
-    To dump or export the TCP/IP configuration, use this command:
-        netsh -c interface dump > PATH_AND_FILENAME.txt
-    Use this command to import the TCP/IP configuration:
-        netsh -f PATH_AND_FILENAME.txt
-
-        netsh -c interface ip dump
-        netsh -c interface ipv6 dump
-
-        Add an address to an interface
-            netsh interface ip add address name="Ethernet 34" address=192.168.168.101 mask=255.255.255.0
-        IPv6 version
-            netsh interface ipv6 add address interface="Ethernet 5" address=fd01::172:31:218:39/64
-
-        Remove an address from an interface
-            netsh interface ip delete address "Ethernet 34" addr=192.168.168.101
-            netsh interface ipv6 delete address interface="Local Area Connection 2" address="fd7e:df1d:94d9:0:381d:a3b4:8849:b4bf"
-     */
-    const {exec} = require('child_process');
-
-    const getIPVersion = (ipVersion) => {
-        let ipVersionCommand = 'ip';
-        if (ipVersion === 6)
-            ipVersionCommand = 'ipv6';
-        return ipVersionCommand;
-    };
-
-    const executeCommand = (command) => {
-        console.debug('Executing command: ' + command);
-        const executedCommand = exec(command, (error, stdout, stderr) => {
-            if (error)
-                console.log(error);
-            // console.log(stdout);
-            // console.log(stderr);
-        });
-
-        executedCommand.on('exit', (code) => {
-            console.debug('Command process exited with exit code: ' + code);
-        })
-    };
-
-    const deleteAdapterAddress = (adapterName, addr, ipVersion=4) => {
-        let ipVersionCommand = getIPVersion(ipVersion);
-        const command = `netsh interface ${ipVersionCommand} delete address "${adapterName}" addr=${addr}`;
-        executeCommand(command);
-    };
-
-    const addAdapterIPv4Address = (adapterName, addr, subnetMask='255.255.255.0') => {
-        let command = `netsh interface ip add address "${adapterName}" address=${addr} mask=${subnetMask}`;
-        executeCommand(command);
-    };
-
-    const addAdapterIPv6Address = (adapterName, cidr) => {
-        const command = `netsh interface ipv6 add address interface="${adapterName}" addr=${cidr}`;
-        executeCommand(command);
-    };
-
     const applyAdapterConfig = (savedAdapters) => {
         for (const adapter of Object.keys(savedAdapters)) {
             let currentAdapterConfig = getNetworkInterfacesFiltered()[adapter];
@@ -146,14 +91,14 @@
                 let ipVersion = 4;
                 if (nic.family === 'IPv6')
                     ipVersion = 6;
-                deleteAdapterAddress(adapter, nic.address, ipVersion);
+                netsh.deleteAdapterAddress(adapter, nic.address, ipVersion);
             }
             // Iterate over saved adapter and add its IP addresses
             for (const nic of savedAdapters[adapter]) {
                 if (nic.family === 'IPv6')
-                    addAdapterIPv6Address(adapter, nic.cidr);
+                    netsh.addAdapterIPv6Address(adapter, nic.cidr);
                 else
-                    addAdapterIPv4Address(adapter, nic.address, nic.netmask);
+                    netsh.addAdapterIPv4Address(adapter, nic.address, nic.netmask);
             }
         }
     };
@@ -238,19 +183,19 @@
                 applyAdapterConfig(this.savedAdapters);
             },
             executeAddIPv4Address: function(event) {
-                addAdapterIPv4Address(this.selectedAddToAdapter, this.addIPv4Address, this.addSubnet);
+                netsh.addAdapterIPv4Address(this.selectedAddToAdapter, this.addIPv4Address, this.addSubnet);
                 this.addIPv4Address = '';
                 this.addSubnet = '';
             },
             executeAddIPv6Address: function(event) {
-                addAdapterIPv6Address(this.selectedAddToAdapter, this.addIPv6Address);
+                netsh.addAdapterIPv6Address(this.selectedAddToAdapter, this.addIPv6Address);
                 this.addIPv6Address = '';
             },
             executeRemoveAddress: function($event, adapterName, nic) {
                 let ipVersion = 4;
                 if (nic.family === 'IPv6')
                     ipVersion = 6;
-                deleteAdapterAddress(adapterName, nic.address, ipVersion);
+                netsh.deleteAdapterAddress(adapterName, nic.address, ipVersion);
             }
         }
     }
